# apps/ledgerline-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ledgerline API",
    "description": "Local statement analyzer. Binds 127.0.0.1 and has no authentication: it is a single-user process holding every statement its owner has imported.",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code"
          },
          "message": {
            "type": "string"
          },
          "rowIndexes": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "description": "Present on `zero_amount_rows`: the rows that parsed to $0.00."
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/imports": {
      "post": {
        "operationId": "uploadImports",
        "summary": "Upload one or more statement files",
        "tags": [
          "imports"
        ],
        "description": "Stages and parses; commits nothing. A byte-identical re-upload returns the existing import untouched (spec 3.3, idempotency layer one).",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "imports": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "import": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "accountId": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "sourceFilename": {
                                "type": "string"
                              },
                              "fileSha256": {
                                "type": "string"
                              },
                              "fileSizeBytes": {
                                "type": "integer"
                              },
                              "formatProfileId": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "periodStart": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "periodEnd": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "rowsParsed": {
                                "type": "integer"
                              },
                              "rowsInserted": {
                                "type": "integer"
                              },
                              "rowsDuplicate": {
                                "type": "integer"
                              },
                              "status": {
                                "type": "string",
                                "enum": [
                                  "uploaded",
                                  "needs_mapping",
                                  "staged",
                                  "committed",
                                  "failed"
                                ]
                              },
                              "parser": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "parserVersion": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "errorDetail": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "diagnosticsJson": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "importedAt": {
                                "type": [
                                  "null",
                                  "string"
                                ]
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "type": "string"
                              }
                            }
                          },
                          "created": {
                            "type": "boolean"
                          },
                          "accountSuggestion": {
                            "type": [
                              "null",
                              "object"
                            ],
                            "properties": {
                              "accountId": {
                                "type": "string"
                              },
                              "reason": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listImports",
        "summary": "Import history",
        "tags": [
          "imports"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "accountId": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "sourceFilename": {
                        "type": "string"
                      },
                      "fileSha256": {
                        "type": "string"
                      },
                      "fileSizeBytes": {
                        "type": "integer"
                      },
                      "formatProfileId": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "periodStart": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "periodEnd": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "rowsParsed": {
                        "type": "integer"
                      },
                      "rowsInserted": {
                        "type": "integer"
                      },
                      "rowsDuplicate": {
                        "type": "integer"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "uploaded",
                          "needs_mapping",
                          "staged",
                          "committed",
                          "failed"
                        ]
                      },
                      "parser": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "parserVersion": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "errorDetail": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "diagnosticsJson": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "importedAt": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/imports/{id}/commit": {
      "post": {
        "operationId": "commitImport",
        "summary": "Commit a staged import",
        "tags": [
          "imports"
        ],
        "description": "Idempotent. Applies the multiset merge rule, then the near-duplicate resolutions, then refund pairing — all inside one transaction, so a partial import never lands (spec 3.3, 2.5).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolutions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "rowIndex",
                        "existingTransactionId",
                        "resolution"
                      ],
                      "properties": {
                        "rowIndex": {
                          "type": "integer"
                        },
                        "existingTransactionId": {
                          "type": "string"
                        },
                        "resolution": {
                          "type": "string",
                          "enum": [
                            "replace",
                            "keep_both",
                            "skip"
                          ]
                        }
                      }
                    }
                  },
                  "allowZeroAmountRows": {
                    "type": "boolean",
                    "description": "Store $0 rows as trial authorizations. Without it a non-pending $0 row is refused as a probable misparse (spec 3.2)."
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "415": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}